[Refactor] Add Typescript
