+++ .gitignore
@@ -12,3 +12,6 @@ lerna-debug.log*
 # Ignore built ts files
 __tests__/runner/*
 lib/**/*
+
+# Secrets
+.secrets
