+++ .nvmrc
@@ -1,1 +1,1 @@
-v10.16.0
+v20.11.0
